requirements.txt
# sagt was wofür ist
discord.py
python-dotenv